Remove Geoffrey
Geoffrey was and is on good terms with the EF;
Following recently shifting attitudes for crypto in Kazakhstan,
his project needed to take a turn away from DAOs
and toward other endeavors that don't fit within the EF's scope
of the fellowship, so he will conclude his fellowship early.
We wish him the best of luck.